requirement.txt
streamlit
zxcvbn-python